dags/requirement.txt
absl-py==1.3.0
alembic==1.8.1
anyio==3.6.2
apache-airflow==2.4.3
apache-airflow-providers-common-sql==1.3.1
apache-airflow-providers-ftp==3.2.0
apache-airflow-providers-http==4.1.0
apache-airflow-providers-imap==3.1.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.0.0
astunparse==1.6.3
attrs==22.1.0
Babel==2.11.0
blinker==1.5
cachelib==0.9.0
cachetools==5.2.0
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
cron-descriptor==1.2.32
croniter==1.3.8
cryptography==38.0.4
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.4
filelock==3.8.0
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.50.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.2
httpx==0.23.1
huggingface-hub==0.11.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.8.0
libclang==14.0.6
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.1
mdurl==0.1.2
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
praw==7.6.1
prawcore==2.3.0
prison==0.2.1
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.2
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==7.0.0
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
rsa==4.9
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
text-unidecode==1.3
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
update-checker==0.18.0
urllib3==1.26.13
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
